# Make the scan of relID optional for GDS algorithms