--- a/src/function/gds/ssp_paths.cpp
+++ b/src/function/gds/ssp_paths.cpp
@@ -21,12 +21,14 @@ public:
     std::vector<nodeID_t> edgeCompute(nodeID_t boundNodeID, graph::NbrScanState::Chunk& resultChunk,
         bool isFwd) override {
         std::vector<nodeID_t> activeNodes;
-        resultChunk.forEach([&](auto nbrNodeID, auto edgeID) {
+        resultChunk.forEach([&](auto neighbors, auto propertyVectors, auto i) {
+            auto nbrNodeID = neighbors[i];
             auto iter = frontierPair->getNextFrontierValue(nbrNodeID.offset);
             if (iter == FRONTIER_UNVISITED) {
                 if (!block->hasSpace()) {
                     block = bfsGraphManager->getCurrentGraph()->addNewBlock();
                 }
+                auto edgeID = propertyVectors[0]->template getValue<nodeID_t>(i);
                 bfsGraphManager->getCurrentGraph()->addSingleParent(frontierPair->getCurrentIter(),
                     boundNodeID, edgeID, nbrNodeID, isFwd, block);
                 activeNodes.push_back(nbrNodeID);