# Description

Under quite a few cases, relID is not needed for computation while we still scan from rel tables. This PR makes the scan of relID optional and configured per algorithm.

Performance comparison between this branch and master on WCC and K-Core are as follows. This is done on a Mac mini with m4 pro chip, 64GB RAM and 1TB SSD.
| Algorithm | Master - cold (ms) | Master - warm (ms) | This branch - cold (ms) | This branch - warm (ms) |
| ---------- | ------------------ | -------------------- | ----------------------- | ------------------------- |
| WCC          | 140.5                       | 92.6                           |  87.1                                | 75.6                                    |
| K-CORE     | 9612.7                     | 9241.8                       | 7168.3                            | 6278.9                                |

The performance different on PageRank is not visible.

TODO:
- [ ] Get rid of the hard coding when accessing rel property vectors inside each algorithm.